requirements.txt
pandas==2.0.3 
sqlalchemy==1.4.46
psycopg2-binary==2.9.6 
streamlit==1.25.0 
python-dotenv==1.0.0 
plotly==5.18.0 
reportlab==4.0.7 
pika==1.3.2 
prometheus-client==0.17.1 
apache-airflow==2.7.3 
mkdocs==1.5.3
scikit-learn==1.2.2